requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
exceptiongroup==1.1.0
frozenlist==1.3.3
grpcio==1.48.1
grpcio-tools==1.48.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
multidict==6.0.4
nest-asyncio==1.5.6
openai==0.26.4
outcome==1.2.0
Pillow==9.4.0
protobuf==3.19.5
PySocks==1.7.1
python-dotenv==0.21.1
python-telegram-bot==20.0
requests==2.28.2
rfc3986==1.5.0
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stability-sdk==0.3.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
undetected-chromedriver==3.2.1
urllib3==1.26.14
websockets==10.4
wsproto==1.2.0
yarl==1.8.2
